Toying with fields_for a has_many association and finding a way to have NestedParams build new records from those generated forms.
In the case of a new record we simply generate a fictive id, but since we don't need it afterwards there's no point in keeping
a counter around, thus we simply generate an id like 'new_12345' where the number is the #object_id of the record.
We can then catch these new records in the NestedParams writer method and build new records.